# [bug#12506] Type-test-override `IndexedSeqOps#foreach`