Effectively override `IndexedSeqOps#foreach` to loop
over indices instead of using an `Iterator` by adding a
type test to `IterableOnceOps#foreach`.

Fixes scala/bug#12506