@NthPortal 2021-12-05
I'll try to get to benchmarking. unfortunately, my server is hosting increasingly many things, making it increasingly unsuited to benchmarking, but hey—it's probably still fine