@lrytz 2021-12-06
I think we have other such "overrides" in place where a real override would be binary incompatible. At least in 2.12 we accumulated a bunch of them over time.