@som-snytt 2021-12-07
The linked commit from strawman overrides all the similar methods such as `forall`.

I think testing is important because it's easy to accidentally deoptimize on the jvm.