@julienrf 2021-12-13
I just ran the following benchmark on my machine:

~~~ scala
package scala.collection

import java.util.concurrent.TimeUnit
import org.openjdk.jmh.annotations._
import org.openjdk.jmh.infra.Blackhole

import scala.collection.immutable.ArraySeq

@BenchmarkMode(Array(Mode.AverageTime))
@Fork(1)
@Threads(1)
@Warmup(iterations = 5)
@Measurement(iterations = 8)
@OutputTimeUnit(TimeUnit.NANOSECONDS)
@State(Scope.Benchmark)
class ForEachBenchmark {

  @Param(Array("0", "1", "10", "1000", "10000", "100000"))
  var size: Int = _
  var array: Array[String] = _
  var vector: Vector[String] = _
  var arraySeq: ArraySeq[String] = _
  var arrayBuffer: mutable.ArrayBuffer[String] = _

  @Setup(Level.Trial) def initNumbers: Unit = {
    array = Array.fill(size)("foo")
    vector = Vector.fill(size)("foo")
    arraySeq = ArraySeq.fill(size)("foo")
    arrayBuffer = mutable.ArrayBuffer.fill(size)("foo")
  }

  @Benchmark def array_foreach(bh: Blackhole): Unit =
    bh.consume(array.foreach(bh.consume))

  @Benchmark def vector_foreach(bh: Blackhole): Unit =
    bh.consume(vector.foreach(bh.consume))

  @Benchmark def arraySeq_foreach(bh: Blackhole): Unit =
    bh.consume(arraySeq.foreach(bh.consume))

  @Benchmark def arrayBuffer_foreach(bh: Blackhole): Unit =
    bh.consume(arrayBuffer.foreach(bh.consume))

}
~~~

Here are the results:

Before:

~~~ text
[info] ForEachBenchmark.arrayBuffer_foreach       0  avgt    8       3.607 ±     0.111  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach       1  avgt    8       8.205 ±     0.217  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach      10  avgt    8      52.256 ±     2.819  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach    1000  avgt    8    4652.913 ±   104.411  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach   10000  avgt    8   47493.288 ±  1164.611  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach  100000  avgt    8  497097.001 ± 21642.157  ns/op
~~~

After:

~~~ text
[info] ForEachBenchmark.arrayBuffer_foreach       0  avgt    8       3.342 ±      0.039  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach       1  avgt    8       7.650 ±      0.593  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach      10  avgt    8      48.327 ±      1.151  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach    1000  avgt    8    4598.888 ±    626.385  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach   10000  avgt    8   48920.076 ±   3978.785  ns/op
[info] ForEachBenchmark.arrayBuffer_foreach  100000  avgt    8  494307.976 ±  51521.829  ns/op
~~~

Before:

~~~ text
[info] ForEachBenchmark.arraySeq_foreach          0  avgt    8       3.909 ±     0.096  ns/op
[info] ForEachBenchmark.arraySeq_foreach          1  avgt    8       7.309 ±     0.168  ns/op
[info] ForEachBenchmark.arraySeq_foreach         10  avgt    8      45.904 ±     0.638  ns/op
[info] ForEachBenchmark.arraySeq_foreach       1000  avgt    8    3455.772 ±   112.868  ns/op
[info] ForEachBenchmark.arraySeq_foreach      10000  avgt    8   34451.301 ±   254.424  ns/op
[info] ForEachBenchmark.arraySeq_foreach     100000  avgt    8  344377.487 ±  3330.170  ns/op
~~~

After:

~~~ text
[info] ForEachBenchmark.arraySeq_foreach          0  avgt    8       4.248 ±      0.652  ns/op
[info] ForEachBenchmark.arraySeq_foreach          1  avgt    8       8.714 ±      0.354  ns/op
[info] ForEachBenchmark.arraySeq_foreach         10  avgt    8      52.733 ±      4.276  ns/op
[info] ForEachBenchmark.arraySeq_foreach       1000  avgt    8    4399.159 ±    868.108  ns/op
[info] ForEachBenchmark.arraySeq_foreach      10000  avgt    8   44252.763 ±   7538.913  ns/op
[info] ForEachBenchmark.arraySeq_foreach     100000  avgt    8  526270.677 ± 611607.069  ns/op
~~~

Before:

~~~ text
[info] ForEachBenchmark.array_foreach             0  avgt    8       3.355 ±     0.051  ns/op
[info] ForEachBenchmark.array_foreach             1  avgt    8       6.762 ±     0.169  ns/op
[info] ForEachBenchmark.array_foreach            10  avgt    8      70.455 ±     0.758  ns/op
[info] ForEachBenchmark.array_foreach          1000  avgt    8    3878.833 ±  3844.431  ns/op
[info] ForEachBenchmark.array_foreach         10000  avgt    8   31033.782 ±   840.745  ns/op
[info] ForEachBenchmark.array_foreach        100000  avgt    8  314350.427 ± 23601.784  ns/op
~~~

After (Arrays should not be affected by the PR):

~~~ text
[info] ForEachBenchmark.array_foreach             0  avgt    8       3.505 ±      0.334  ns/op
[info] ForEachBenchmark.array_foreach             1  avgt    8       7.555 ±      0.797  ns/op
[info] ForEachBenchmark.array_foreach            10  avgt    8      38.023 ±      5.219  ns/op
[info] ForEachBenchmark.array_foreach          1000  avgt    8    3358.272 ±    408.329  ns/op
[info] ForEachBenchmark.array_foreach         10000  avgt    8   35310.950 ±   6418.211  ns/op
[info] ForEachBenchmark.array_foreach        100000  avgt    8  367379.432 ±  45226.684  ns/op
~~~

Before:

~~~ text
[info] ForEachBenchmark.vector_foreach            0  avgt    8       3.366 ±     0.127  ns/op
[info] ForEachBenchmark.vector_foreach            1  avgt    8       7.075 ±     0.180  ns/op
[info] ForEachBenchmark.vector_foreach           10  avgt    8      35.302 ±     0.917  ns/op
[info] ForEachBenchmark.vector_foreach         1000  avgt    8    3763.463 ±   244.501  ns/op
[info] ForEachBenchmark.vector_foreach        10000  avgt    8   37467.608 ±   569.759  ns/op
[info] ForEachBenchmark.vector_foreach       100000  avgt    8  384386.185 ± 29821.773  ns/op
~~~

After:

~~~ text
[info] ForEachBenchmark.vector_foreach            0  avgt    8       3.668 ±      0.558  ns/op
[info] ForEachBenchmark.vector_foreach            1  avgt    8       7.121 ±      0.155  ns/op
[info] ForEachBenchmark.vector_foreach           10  avgt    8      36.102 ±      3.690  ns/op
[info] ForEachBenchmark.vector_foreach         1000  avgt    8    3977.236 ±    561.623  ns/op
[info] ForEachBenchmark.vector_foreach        10000  avgt    8   37354.477 ±   1814.811  ns/op
[info] ForEachBenchmark.vector_foreach       100000  avgt    8  375880.352 ±  17324.349  ns/op
~~~

Not sure if I did something wrong. But there seems to be no impact on the performance.