@Ichoran 2021-12-13
If there's no difference either way for anything, is it actually worth adding to the code complexity?  Nominally it avoids an object creation, which one would think was good, but it looks to me like the object creation is getting optimized out or you'd be able to see the initializer cost.

Can you check GC churn before/after to verify?

I'd imagine in cases with less optimization, you'd get increased memory usage and initializer costs, but then you'd get a `Vector` slowdown because it's not very good at indexing.