@SethTisue 2021-12-13
@WojciechMazur you were the original reporter on Discord, I believe. Could you test this PR with your code, or supply your code in benchmark form?